Extract test case from GroupService to GroupTest
